fix(test): update NewTestDB to work with server DBMSs
Update NewTestDB() to call ResetModels() as part of test DB
initialization. This ensures tests start from a clean slate when running
on server-based DBMSs that may persist results from prior tests
otherwise.

As this requires access to ResetModels(), NewTestDB() has been moved to
pkg/model for pkg/test (the latter is removed as NewTestDB was its only
contents).
